Changes to enable robot code to compile with 2018beta FRC libraries.
Most of the changes revolve around changing RobotDrive to
DifferentialDrive, and changing CANTalon to TalonSRX.  The
CTRE libraries for the Talons are in a state of flux and do
not explicitly implement the SpeedController interface (although
all the methods are actually implemented).  So, the MyTalonSRX
class was created to tag the SpeedController interface.